Test longitudinal measurement invariance

Test longitudinal measurement invariance

long_minvariance(
  data,
  n_timepoints,
  n_items,
  add = NULL,
  remove = NULL,
  measure_name = "x",
  time_str = "_t",
  item_str = "_i",
  details = FALSE,
  return = c("fit_statistics", "model_tests", "lavaan_objects", "lavaan_syntax")
)

Arguments

data
Data in wide format (one row for each participant, one column for each item at each time point)

n_timepoints
Number of time points

n_items
Number of items

add
lavaan syntax to add to the model

remove
List, remove specific parts of the automatically generated model

measure_name
String, name of the measure

time_str
String, prefix of the point specifier

item_str
String, prefix of the item specifier

details
Return details of selected output selected in return argument

return
What should be returned ("fit_statistics", "model_tests", "lavaan_objects", "lavaan_syntax")

Value